Formata colunas numericas em pt-BR

Atalho para gt::fmt_number() com separador de milhar ponto e decimal virgula, que e o que os artigos usam em toda tabela.

Usage

fmt_br(g, colunas, decimais = 0)

Arguments

g

Objeto gt_tbl.

colunas

Colunas a formatar (tidyselect).

decimais

Casas decimais.

Value

O gt_tbl com as colunas formatadas.

See also

Examples

fmt_br(tabela_gt(head(mtcars[, 1:3])), disp, decimais = 1)

  

mpg
      	cyl
      	disp
    

21.0
	6
	160,0

21.0
	6
	160,0

22.8
	4
	108,0

21.4
	6
	258,0

18.7
	8
	360,0

18.1
	6
	225,0